+++ R/Tmisc-package.R
@@ -1,12 +1,8 @@
-#' Tmisc
-#' 
-#' Stephen Turner's miscellaneous functions
-#'
-#' @author Stephen Turner
-#'
-#' @name Tmisc
-#' @docType package
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
 NULL
-## quiets concerns of R CMD check re: the .'s that appear in pipelines
-## and the "n" that is produced by dplyr::count() in a pipeline
+
 if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", "n"))